+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+  - Use iscript animations from a source file instead of a binary
   - Add a new trigger action: "Execute Lua"
     * Implement a `Scenario` type
     * Implement a `Player` type